+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Dockerfile for building container. [#19](https://github.com/IN-CORE/maestro-service/issues/19)
 - GitHub action for automation of container building. [#23](https://github.com/IN-CORE/maestro-service/issues/23)
+- Add User in Steps model to record last updated status. [#20](https://github.com/IN-CORE/maestro-service/issues/20)
 
 ### Change
 - API takes the testbed name in its routes. [#31](https://github.com/IN-CORE/maestro-service/issues/31)